% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{aggregate_major_companies}
\alias{aggregate_major_companies}
\title{Collapes names of major companies}
\usage{
aggregate_major_companies(names_column)
}
\arguments{
\item{names_column}{a string vector of company names}
}
\value{
the string vector with and aggregation of the names of major companies
}
\description{
In a vector of company names the same company usually appears with various names.
This function lists some of them and aggregates them unde one main name.
}
