% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_tjmt}
\alias{filter_tjmt}
\title{Filter only consumer lawsuits}
\usage{
filter_tjmt(loaded_tjmt)
}
\arguments{
\item{loaded_tjmt}{data_frame obtained from load_tjmt}
}
\description{
Filter only consumer lawsuits
}
