% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fix_nomes}
\alias{fix_nomes}
\title{Fix column of brazillian company names}
\usage{
fix_nomes(names_column)
}
\arguments{
\item{names_column}{a string vector of company names}
}
\value{
a better version of names_column
}
\description{
This function receives a string vector of company names,
substitutes all types of "s/a" by "@s@a@", all types of "ltda." by "@l@t@d@a" and all types of "m.e." by "@m@e@".
At the end, it trims the vector, replaces multiple spaces by one and removes accentuation
}
