% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_tj}
\alias{load_tj}
\alias{load_tjrs}
\alias{load_tjsp}
\alias{load_tjmt}
\title{Carrega os dados brutos do Tribunal para iniciar limpeza}
\usage{
load_tj(raw_data, tj = "tjsp")

load_tjrs(raw_data)

load_tjsp(raw_data)

load_tjmt(assuntos, processos, partes)
}
\arguments{
\item{raw_data}{data.frame lido diretamente dos raw_data passados pelo
Tribunal.}

\item{tj}{nome do Tribunal. Pode ser \code{"tjrs"}, \code{"tjsp"},
\code{"tjms"} ou \code{"tjba"}.}
}
\value{
tibble com as colunas \itemize{ \item{"n_processo"}{Número do
  processo} \item{"dt_dist"}{Data de distribuição} \item{"classe"}{Classe}
  \item{"assunto"}{Assunto} \item{"comarca"}{Comarca} \item{"foro"}{Foro}
  \item{"autor"}{Autor} \item{"reu"}{Réu} \item{"vl_causa"}{Valor da causa} }
}
\description{
Carrega os dados brutos do Tribunal para iniciar limpeza. Basicamente faz a
seleção das colunas de interesse para essa fase da pesquisa e arruma os
formatos para ficar tudo igual.
}
